refactor compareAddresses
